Bump BootStrap and FontAwesome
Bumped Bootstrap to the latest version. Also removed the old, built-in version of FontAwesome in favor of serving the 5.14 Pro version from a kit of mine.  This is faster as it only serves the icons we use.  Also removed a reference in the footer that was manually declaring a jQuery version.  This is bad WordPress practice as declaring a conflicting version from Wordpress Core can break functionality.