swap the submodule for composer loaded one.
The SDK for authorize.net requires a specific folder structure to exist even when not loaded from Composer.  I don't want future maintainers to have to know how to use Composer so I am loading it in and committing all the files to our source code.  The submodule method could have worked, but still required the "vendor" folder structure and specific build files. The documentation suggested building the sdk from a different device and transferring it.